fix: Change the main entry point to dist/src/index.js chore: Have the samples depend on the SDK as '@google/genai'. test: Build the samples on CI.
I'm keeping it dist/src/index.js as that's where tsc puts it right now.
We'll be moving to bundled files soon and will get rid of the src/ subpath with that.

I turned the `samples/` folder into a separate module so it can depend on the SDK the same way as external users.

PiperOrigin-RevId: 729262770