chore: Configure and test the published tarball. chore: Silence api-extractor tsdoc warnings.
This adds a `files` entry to package.json to control the files that are included in the tarball published to npm.
And adds a test that builds the tarball locally and makes sure the samples project can build against the tarball.

This change also silences the api-extractor tsdoc warnings we were getting, we don't care for them as we're using typedoc and not tsdoc, and we need them disabled so that the prod build completes with a 0 exit code.

PiperOrigin-RevId: 731019749