docs: Expose the cross platform client at the root, hide the node and web submodules.
All 3 clients (web, node, and the crossplatform) should have the same API surface and same documentation,
to make it easier to navigate the docs site I suggest we keep the portable client at the root, use it for the docsite, and hide the node and web submodules.

I'm splitting the actual docs regeneration to the next change to keep the diff clear.

Before:

![The docsite before this change](https://github.com/user-attachments/assets/dc4b13fe-b9e5-4ead-8892-464d65b0a89e)

After:

![The docsite after this change](https://github.com/user-attachments/assets/88e34569-5b58-47b8-a39d-f5857192ccd6)

In this change:
  1. typedoc.json is updated to exclude everything under web/* and node*
  2. cross/_cross_platform_client.ts is moved to client.ts
  3. affected referenced are updated

PiperOrigin-RevId: 732999430