fix: Use authentication headers as provided by google-auth-library
We used to explicitly set the 'Authorization:' header, which was enough for most cases.
However the x-goog-user-project header is needed for some calls.
Instead of maintaining the knowledge of which authentication headers to set in this package
this change makes it so we rely on the header information that is coded in the google-auth-library

PiperOrigin-RevId: 742867644